test(onboard): isolate deployment messaging coverage
